django_project/django_project/requirements.txt
amqp==2.5.0
anyjson==0.3.3
billiard==3.6.0.0
celery==4.3.0
chardet==2.0.1
Django==1.11.17
django-crispy-forms==1.7.2
kombu==4.6.3
numpy==1.15.4
Pillow==5.4.1
pycurl==7.19.3
python-apt===0.9.3.5ubuntu3
pytz==2018.7
requests==2.2.1
six==1.5.2
ssh-import-id==3.21
urllib3==1.7.1
vine==1.3.0
